Added docs

+++ dspace-api/src/main/java/org/dspace/content/clarin/ClarinLicenseLabelServiceImpl.java
@@ -1,3 +1,10 @@
+/**
+ * The contents of this file are subject to the license and copyright
+ * detailed in the LICENSE and NOTICE files at the root of the source
+ * tree and available online at
+ *
+ * http://www.dspace.org/license/
+ */
 package org.dspace.content.clarin;
 
 import org.apache.commons.lang.NullArgumentException;
@@ -16,6 +23,13 @@ import java.sql.SQLException;
 import java.util.List;
 import java.util.Objects;
 
+/**
+ * Service implementation for the Clarin License Label object.
+ * This class is responsible for all business logic calls for the Clarin License Label object and is autowired by spring.
+ * This class should never be accessed directly.
+ *
+ * @author Milan Majchrak (milan.majchrak at dataquest.sk)
+ */
 public class ClarinLicenseLabelServiceImpl implements ClarinLicenseLabelService {
 
     private static final Logger log = LoggerFactory.getLogger(ClarinLicenseServiceImpl.class);
@@ -58,17 +72,32 @@ public class ClarinLicenseLabelServiceImpl implements ClarinLicenseLabelService 
     }
 
     @Override
-    public List<ClarinLicenseLabel> findAll(Context context) throws SQLException {
+    public List<ClarinLicenseLabel> findAll(Context context) throws SQLException, AuthorizeException {
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "You must be an admin to create an Clarin License Label");
+        }
+
         return clarinLicenseLabelDAO.findAll(context, ClarinLicenseLabel.class);
     }
 
     @Override
-    public void delete(Context context, ClarinLicenseLabel license) throws SQLException {
+    public void delete(Context context, ClarinLicenseLabel license) throws SQLException, AuthorizeException {
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "You must be an admin to create an Clarin License Label");
+        }
+
         clarinLicenseLabelDAO.delete(context, license);
     }
 
     @Override
-    public void update(Context context, ClarinLicenseLabel newClarinLicenseLabel) throws SQLException {
+    public void update(Context context, ClarinLicenseLabel newClarinLicenseLabel) throws SQLException, AuthorizeException {
+        if (!authorizeService.isAdmin(context)) {
+            throw new AuthorizeException(
+                    "You must be an admin to create an Clarin License Label");
+        }
+
         if (Objects.isNull(newClarinLicenseLabel)) {
             throw new NullArgumentException("Cannot update licenseLabel because the clarinLicenseLabel is null");
         }
